# openapi.yaml
openapi: 3.0.3
info:
  title: Ticket Booking System API
  version: 1.0.0
  description: |
    Ticket Booking System with JWT authentication, concurrency-safe reservations,
    and automatic order expiration (15 minutes).

    Business Rules:
    - Password is stored as hashed value (never returned in API)
    - Prevent duplicate events (same venue + datetime)
    - Prevent double booking using atomic reservation
    - Orders expire after 15 minutes if still PENDING
    - Expired orders are CANCELLED and tickets are released to AVAILABLE

servers:
  - url: http://localhost:3000/api

tags:
  - name: Auth
  - name: Events
  - name: Venues
  - name: Tickets
  - name: Orders

paths:

  /auth/register:
    post:
      tags: [Auth]
      summary: Register user (email + password)
      description: Password will be hashed before storing in database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and return JWT token
      description: Validate email and hashed password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /venues:
    post:
      tags: [Venues]
      summary: Create venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVenueRequest'
      responses:
        '201':
          description: Venue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'

    get:
      tags: [Venues]
      summary: Get all venues
      responses:
        '200':
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Venue'

  /events:
    post:
      tags: [Events]
      summary: Create event
      description: Prevent duplicate event (same venue + datetime)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'

    get:
      tags: [Events]
      summary: Get all events
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'

  /tickets:
    get:
      tags: [Tickets]
      summary: Get tickets
      parameters:
        - name: eventId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'

  /orders:
    post:
      tags: [Orders]
      summary: Create order (reserve tickets)
      description: |
        - Requires JWT authentication
        - Prevent race condition using atomic update (e.g. SELECT FOR UPDATE)
        - Tickets move from AVAILABLE → RESERVED
        - Order expires in 15 minutes if not confirmed
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created (PENDING)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

    get:
      tags: [Orders]
      summary: Get user orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /orders/{id}/confirm:
    patch:
      tags: [Orders]
      summary: Confirm order
      description: |
        - Requires JWT
        - Change order status to CONFIRMED
        - Tickets RESERVED → SOLD
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /orders/{id}/cancel:
    patch:
      tags: [Orders]
      summary: Cancel order
      description: |
        - Requires JWT
        - Change order status to CANCELLED
        - Tickets RESERVED → AVAILABLE
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    User:
      type: object
      description: User response (password never exposed)
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time

    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
          minLength: 6

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true

    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/User'

    Venue:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateVenueRequest:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        location:
          type: string

    Event:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        venueId:
          type: string
        datetime:
          type: string
          format: date-time
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time

    CreateEventRequest:
      type: object
      required: [name, venueId, datetime]
      properties:
        name:
          type: string
        venueId:
          type: string
        datetime:
          type: string
          format: date-time

    Ticket:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        seatNumber:
          type: string
        status:
          type: string
          enum: [AVAILABLE, RESERVED, SOLD]

    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        ticketIds:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    CreateOrderRequest:
      type: object
      required: [ticketIds]
      properties:
        ticketIds:
          type: array
          items:
            type: string